requirements.txt
python-telegram-bot==21.3
python-dotenv==1.0.1

# Google Sheets + Drive logging
gspread==6.1.2
google-auth==2.29.0
google-api-python-client==2.126.0
